Refactor: error classes in utils
