+++ src/main/java/com/jogamais/ufcg/utils/CourtError.java
@@ -1,4 +1,4 @@
-package com.jogamais.ufcg.util;
+package com.jogamais.ufcg.utils;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
