+++ src/main/java/com/jogamais/ufcg/utils/CustomTypeError.java
@@ -1,13 +1,9 @@
-package com.jogamais.ufcg.util;
+package com.jogamais.ufcg.utils;
 
 public class CustomTypeError {
     private String errorMessage;
 
     public CustomTypeError(String errorMessage) {
         this.errorMessage = errorMessage;
-    }
-
-    public String getErrorMessage() {
-        return  errorMessage;
     }
 }
