+++ src/main/java/com/jogamais/ufcg/utils/UserError.java
@@ -1,11 +1,11 @@
-package com.jogamais.ufcg.util;
+package com.jogamais.ufcg.utils;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 public class UserError {
 
-    static final String USER_NOT_EXIST = "O usuário não cadastrado.";
+    static final String USER_NOT_EXIST = "Usuário(s) não cadastrado(s).";
 
     static final String USER_ALREADY_EXIST = "O usuário %s já está cadastrado com o CPF: %s";
 
@@ -16,4 +16,5 @@ public class UserError {
     public static ResponseEntity<CustomTypeError> errorUserAlreadyExist(String cpf) {
         return new ResponseEntity<CustomTypeError>(new CustomTypeError(String.format(UserError.USER_ALREADY_EXIST, cpf)), HttpStatus.CONFLICT);
     }
+
 }
